requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
PyJWT
httpx
cryptography
openai
